fix: repeat history pid para

+++ User/utils/pid.c
@@ -39,7 +39,7 @@ void pid_init(Pid *pid, float kp, float ki, float kd,
     pid->data[i] = 0;
 }
 
-int pid_compute(Pid *pid, short target, short current) {
+int pid_compute(Pid *pid, int target, int current) {
   short err = target - current;
 
   /*
