+++ User/utils/pid.h
@@ -82,6 +82,6 @@ void pid_init(Pid *pid, float kp, float ki, float kd,
  * }
  * ```
  */
-int pid_compute(Pid *pid, short target, short current);
+int pid_compute(Pid *pid, int target, int current);
 
 #endif // !__PID_H__
